feat(runner): move `Columns` class
