+++ crunch/runner/__init__.py
@@ -1,3 +1,5 @@
 import os
 
+from .columns import Columns
+
 is_inside = os.getenv("CRUNCH_INSIDE_RUNNER", "false").lower() == "true"
